shut msvc up
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@9058 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/asr_tts/mod_pocketsphinx/mod_pocketsphinx.c
@@ -363,7 +363,7 @@ static switch_status_t pocketsphinx_asr_get_results(switch_asr_handle_t *ah, cha
 		switch_clear_flag(ps, PSFLAG_HAS_TEXT);
 		
 		conf = ps_get_prob(ps->ps, &ps->uttid);
-		lconf = logmath_log_to_ln(ps_get_logmath(ps->ps), conf);
+		lconf = (int32_t)logmath_log_to_ln(ps_get_logmath(ps->ps), conf);
 		ps->confidence = lconf - lconf - lconf;
 
 		if (ps->confidence > 100) {
